Refactor: styled code python

+++ tester/app/models/Record.py
@@ -6,7 +6,6 @@ from app.schemas.enums.Status import Status
 class Record(BaseModel):
     total: int
     passed: int
-    status_list : list[Status]
-    time_response : list[float]
-    memory_used : list[int]
-
+    status_list: list[Status]
+    time_response: list[float]
+    memory_used: list[float]
